+++ src/beta/chat/index.ts
@@ -1,0 +1,6 @@
+/*
+*   Muna
+*   Copyright © 2025 NatML Inc. All Rights Reserved.
+*/
+
+export { ChatService } from "./chat"