+++ src/beta/types.ts
@@ -1,0 +1,7 @@
+/*
+*   Muna
+*   Copyright © 2025 NatML Inc. All Rights Reserved.
+*/
+
+export type { CreateRemotePredictionInput, RemoteAcceleration } from "./remote"
+export * from "./chat/types"