+++ src/index.ts
@@ -3,7 +3,6 @@
 *   Copyright © 2025 NatML Inc. All Rights Reserved.
 */
 
-export type { CreateRemotePredictionInput, RemoteAcceleration } from "./beta"
 export type { MunaAPIError } from "./client"
 export * from "./muna"
 export * from "./types"